chore: :bug: fix type inference issue in @ioc:Adonis/Lucid/Orm module
The type inference issue in the @ioc:Adonis/Lucid/Orm module has been fixed. This problem was causing unexpected behavior when trying to infer the types of properties within a Lucid ORM model.

The fix involved updating the helper types and recursive functions used in the module to correctly infer the types of properties within a model. Additionally, the ability to exclude properties from the inference process was added by passing the keys of the properties as a second parameter to the InferTypeModel type.

This change ensures type safety and improves the documentation of the model's properties, thus avoiding any confusion or unexpected behavior.